Player Select menu complete
May re-do the text to say "Player 1 press x" instead of what it is now.
Still need backgrounds for the characters - or an overall background.
Fixed logic to enable players during play mode due to moving a script for programming convienience.
Next need to tell players they are out if they get caught petting, and then withdrawl hands animation so only active players are on the screen.